Fix data types for mNormalMapFormat in D3DTX_V6 and D3DTX_V7

+++ DDS_D3DTX_Converter_GUI/DDS_D3DTX_Converter/TelltaleD3DTX/D3DTX_V6.cs
@@ -122,7 +122,7 @@ namespace D3DTX_Converter.TelltaleD3DTX
         /// <summary>
         /// [4 bytes] Defines the format of the normal map.
         /// </summary>
-        public float mNormalMapFormat { get; set; }
+        public int mNormalMapFormat { get; set; }
 
         /// <summary>
         /// [4 bytes] Defines the brightness scale of the texture. (used for lightmaps)
